installment_amount type changed to integer;

+++ db/migrate/20201028140609_create_solicitations.rb
@@ -3,7 +3,7 @@ class CreateSolicitations < ActiveRecord::Migration[6.0]
     create_table :solicitations do |t|
       t.integer :loan_amount
       t.integer :number_of_installments
-      t.float :installment_amount
+      t.integer :installment_amount
       t.boolean :pending, default: true
       t.references :user, null: false, foreign_key: true
